create searchBar component
