+++ frontend/src/components/Topbar/Topbar.jsx
@@ -1,32 +1,22 @@
 import { Box, IconButton, useTheme } from "@mui/material";
 import { useContext } from "react";
-import InputBase from "@mui/material/InputBase";
 import LightModeOutlinedIcon from "@mui/icons-material/LightModeOutlined";
 import DarkModeOutlinedIcon from "@mui/icons-material/DarkModeOutlined";
 import NotificationsOutlinedIcon from "@mui/icons-material/NotificationsOutlined";
 import SettingsOutlinedIcon from "@mui/icons-material/SettingsOutlined";
 import PersonOutlinedIcon from "@mui/icons-material/PersonOutlined";
-import SearchIcon from "@mui/icons-material/Search";
-import { ColorModeContext, tokens } from "../../theme";
+import SearchBar from "../SearchBar/SearchBar";
+
+import { ColorModeContext } from "../../theme";
 
 function Topbar() {
   const theme = useTheme();
-  const colors = tokens(theme.palette.mode);
   const { toggleColorMode } = useContext(ColorModeContext);
 
   return (
     <Box display="flex" justifyContent="space-between" p={2}>
       {/* SEARCH BAR */}
-      <Box
-        display="flex"
-        backgroundColor={colors.primary[400]}
-        borderRadius="3px"
-      >
-        <InputBase sx={{ ml: 2, flex: 1 }} placeholder="Search" />
-        <IconButton type="button" sx={{ p: 1 }}>
-          <SearchIcon />
-        </IconButton>
-      </Box>
+      <SearchBar />
 
       {/* ICONS */}
       <Box display="flex">
